+++ src/main/java/online/kingdomkeys/kingdomkeys/datagen/init/BlockStates.java
@@ -233,7 +233,7 @@ public class BlockStates extends BlockStateProvider {
 			} else if (block instanceof KKOreBlock && name.endsWith("_n")) {
 				simpleNetherOre(itemRegistryObject);
 			} else if (block instanceof GummiBlockBase || name.contains("gummi_meteor")) {
-				simpleBlock(block, new ModelFile.UncheckedModelFile(ResourceLocation.fromNamespaceAndPath(KingdomKeys.MODID, "block/"+name)));
+				simpleBlock(block, new ModelFile.UncheckedModelFile(ResourceLocation.fromNamespaceAndPath(KingdomKeys.MODID, "block/" + name)));
 			} else {
 				simpleBlock(itemRegistryObject);
 			}